% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflFantasyLabsAll.R
\name{fullNFLModelFromCSV}
\alias{fullNFLModelFromCSV}
\title{Full NFL models from CSV function}
\usage{
fullNFLModelFromCSV(week)
}
\arguments{
\item{week}{}
}
\value{
nfl data frame labs model(FULL)
}
\description{
Full NFL models from CSV function
}
\examples{
fullNFLModelFromCSV(week)
}

