% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmRSeleniumFunction.R
\name{getBBMCookies}
\alias{getBBMCookies}
\title{getBBMCookies}
\usage{
getBBMCookies(dockerName = "monsterLab")
}
\value{
all BBM Cookies
}
\description{
getBBMCookies
}
\examples{
getBBMCookies()
}

