% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBBMModelFunction.R
\name{getBBMModel}
\alias{getBBMModel}
\title{Get NBA BBMonster data for Daily Projections}
\usage{
getBBMModel(bbmMonth = "1", bbmDay = "29", bbmYear = "2017",
  bbmCookie = NULL)
}
\arguments{
\item{bbmMonth}{"1"}

\item{bbmDay}{"27"}

\item{bbmYear}{"2017"}

\item{bbmCookie}{bbmonster session __utmz cookie value i.e: "49390133.1485608109.1.1"}
}
\value{
BBMonster data df/csv
}
\description{
Get NBA BBMonster data for Daily Projections
}
\examples{
getBBMModel(bbmMonth="1",bbmDay="20",bbmYear="2017",bbmCookie="49390133.1485608109.1.1")
}

