% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMlbModelFunction.R
\name{getMlbModel}
\alias{getMlbModel}
\title{MLB FantasyLabs Model Retrieve}
\usage{
getMlbModel(modelDate = NULL, cookie = labCookies, View = NULL)
}
\arguments{
\item{modelDate}{"4_2_2017"}

\item{cookie}{labCookies}

\item{View}{H or P}
}
\value{
hitter, pitcher, or ownership percentage models
}
\description{
MLB FantasyLabs Model Retrieve
}
\examples{
getMlbModel("4_2_2017",labCookies,"H")
}

