% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBAmodelandProject.R
\name{getNbaModelandProject}
\alias{getNbaModelandProject}
\title{NBA model Retrieve and projections export}
\usage{
getNbaModelandProject(modelDate = "10_30_2015")
}
\arguments{
\item{modelDate}{}
}
\value{
FantasyLabs NBA Model projections
}
\description{
NBA model Retrieve and projections export
}
\examples{
getNbaModelandProject()
}
