% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaModelRetrieve.R
\name{getNBAPlayerModel}
\alias{getNBAPlayerModel}
\title{NBA Model Retrieve}
\usage{
getNBAPlayerModel(modelDate = NULL, modelSite = "Draftkings",
  slate = NULL, exclude = NULL, cookie = labsC)
}
\arguments{
\item{modelDate}{"1_1_2017"}

\item{modelSite}{Draftkings or FanDuel}

\item{slate}{Main,turbo,late,etc}

\item{exclude}{c("ATL","NYK")}

\item{cookie}{for cURL from FLabs (Chrome Browser)..}
}
\value{
Fantasylabs NBA model
}
\description{
NBA Model Retrieve
}
\examples{
getNBAPlayerModel()

}

