% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaModelRetrieveFanDuel.R
\name{getNBAPlayerModelFD}
\alias{getNBAPlayerModelFD}
\title{NBA Model Retrieve(FanDuel)}
\usage{
getNBAPlayerModelFD(modelDate = NULL, slate = NULL, exclude = NULL,
  cookie = "__cfduid=d8a77c25dbef815632fdbf4f45e81df8b1483248089; LD_U=http\%3A\%2F\%2Fwww.fantasylabs.com\%2F; flid=HQAMnfjSB0ytEUbUy1F_3w; __distillery=13f6068_5d3aa9eb-3e25-403f-be81-5a58b0d371e7-0ef1805cc-ee4b27f5fbc2-cf6a; LD_S=1484395001153; LD_R=; _gat=1; __zlcmid=eNg7LAWTPEIN0T; .AspNet.Cookies=HvlkmjkB2hOLIjpz6d0F83ga6mSltxFn8SlRcPAeSTYlVEK-KGeBLODL74KcuXggUxjSHMhaVKsj105PJl94rsAvzs6SiqsqB3GevubDHupyP26duElDEXAVnqdfW9hUGPmtq7JTr0gMncjlN-vvbQgmSnVT1pg_WxhtojndN5qmtBYK1ZGWOHsvqCapHd7AgFYGwdOfzez01eKVyAtLG0IUXSZnjfXUviFn1U2XeW1VypD_42uykPSBpVyBJ3h1cwSUiGVfCn22BYxejNKo-v--nqwob_7fj2LXNaK3FMBg0qBerQ911pp25SyrhshdxElQf3b4b83Q8SquwW-uuxWhyLj-HL7zk7DblWtK9L-QYk7DEe7w3UoDtrkAG9TiKEbEnBJxv-aI2jmIC9xghrIaBHsDXvq_tnrASbMb4cF0nUmrgZQV3ZEncc0qpHx4FMW-alc8izQJxWQuNy4tNc5hvYgON6oQeOkrSigFjTY; _ga=GA1.2.2034235139.1483248092; LD_T=3afe007d-48de-4954-9c64-a6dd5f392768'")
}
\arguments{
\item{modelDate}{"1_1_2017"}

\item{cookie}{for cURL from FLabs (Chrome Browser)..}
}
\value{
Fantasylabs NBA model
}
\description{
NBA Model Retrieve(FanDuel)
}
\examples{
getNBAPlayerModelFD()

}

