% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNbaPlayerModel_OtherFunction.R
\name{getNbaPlayerModels_Other}
\alias{getNbaPlayerModels_Other}
\title{Get NBA Labs Model for a model otber than 'Finding Value' Model}
\usage{
getNbaPlayerModels_Other(modelDate = "12_18_2016", other = "779922",
  mName = "Phan")
}
\arguments{
\item{modelDate}{"1_1_2017"}

\item{other}{"705700"}

\item{mName}{"Russell_and_Flow"}
}
\value{
newPlayerModelof Desired choice
}
\description{
Get NBA Labs Model for a model otber than 'Finding Value' Model
}
\examples{
getNbaPlayerModels_Other(modelDate="12_18_2016",other = "705700",mName = "Russell_and_Flow")
}

