% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFLBasicLabsModelRetrieve.R
\name{getNflLabModel}
\alias{getNflLabModel}
\title{NFL FantasyLabs Model Retrieve}
\usage{
getNflLabModel(modelYear = "2016", modelWeek = NULL, model = NULL,
  showModelList = FALSE, PositionID = NULL,
  cookie = "__cfduid=d19da73f21ae9852e137438d8ffd398371475796569; LD_U=http\%3A\%2F\%2Fwww.fantasylabs.com\%2F; __distillery=a88774e_344d8f47-a8f7-4094-81ba-d3336d4c2d4d-ba08c6bef-19ba082fafba-d101; LD_R=http\%3A\%2F\%2Fwww.fantasylabs.com\%2Fnfl\%2Fplayer-models\%2F; .AspNet.Cookies=I3uyNArip8P2mwm4iZUdp1aswKJvJUJsgXxxhq2yU-UayOPc51mQDfzLkb2HihonPOWkz3ZPkcBeFOk1SeaGwDn5-Oh61Xu1cUsarqZkutXgbTBwc2V6ITx_u4lMXjuodLaUjd3x0DcSg2Y7YFMQwjduY9bLFWkLyDqr5rutPNS5bxAxO80joxoa7tfhKOC0vLGmO3ZCDMTWGNXAMqiGpfG5p7eiS7drDO5q9AEwYYux_fLWATa04cZXGRkHc3eae_7B_eaGUaFPNqom5dZ2ewqZRHktUEGo4JavxHfv4TMxS4t9URXN3YxK6YGh7edIh1pFHuZ1RQpVolIHQzbuedjxPb_dkSwZ1iVPcplBM10TV3g3NRDZXZVQ0LYWbkFUdqEvbMeiBMos-kNeurxSrlKdwbQTro0vhq_ckdsRFtduXzQ-zxsbcwWHiUrNNOshYTZMau5j2tZzPsPyCOLGrPXJPJd-ZaKSJNIL_gY5pz8; flid=yijf7wz4JEegj3XR4O5vLw; _gat=1; LD_S=1478649158829; __zlcmid=cyfuDnxjy8XwIK; LD_T=6a763f78-d230-4827-f05d-37ca01630a5d; _ga=GA1.2.407536382.1475796573'")
}
\arguments{
\item{modelYear}{(2015,2016)}

\item{modelWeek}{(1-17)}

\item{model}{(1...42)}

\item{showModelList}{default FALSE(takes thru interactive modeling if TRUE)}

\item{PositionID}{(QB.RB.WR,TE,DST)}

\item{cookie}{cookies from browser needed to get into fantasylabs account}
}
\value{
FantasyLabs NFL model
}
\description{
NFL FantasyLabs Model Retrieve
}
\examples{
getNflLabModel(modelYear = "2015",modelWeek = NULL,model = NULL,mod = "", PositionID = NULL)
}
