% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPgaModelFunction.R
\name{getPgaModel}
\alias{getPgaModel}
\title{PGA FantasyLabs Model Retrieve}
\usage{
getPgaModel(modelDate = NULL, cookie = labCookies)
}
\arguments{
\item{modelDate}{"4_2_2017"}

\item{cookie}{labCookies}
}
\value{
Fantasylabs PGA player Model
}
\description{
PGA FantasyLabs Model Retrieve
}
\examples{
getpgaModel("4_2_2017",labCookies)
}

