% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmRSquaredFunction.R
\name{glmRSquared}
\alias{glmRSquared}
\title{GLM regression correlation r^2 value}
\usage{
glmRSquared(fitModel = NULL)
}
\arguments{
\item{fitModel}{glm Model Name i.e:nYearGLM3}
}
\value{
r^2 correlation of regression model
}
\description{
GLM regression correlation r^2 value
}
\examples{
glmRSquared(fitModel=nYearGLM3)
}

