% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterateAllNfl.R
\name{iterateAllNFL}
\alias{iterateAllNFL}
\title{NFL Iterate All Function}
\usage{
iterateAllNFL(weeksFor = NULL, forPosition = "QB")
}
\arguments{
\item{weeksFor}{}

\item{forPosition}{}
}
\value{
AllModelforPosition
}
\description{
NFL Iterate All Function
}
\examples{
iterateAllNFL(weeksFor =NULL,forPosition = "QB")
}
