% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labsRegressionXFunction.R
\name{labsRegressionX}
\alias{labsRegressionX}
\title{Labs Regression X}
\usage{
labsRegressionX(labModel = NULL, fromReg = NULL)
}
\arguments{
\item{labModel}{labs Model to fit data (X)}
}
\value{
labsModel prepared for linear regression fitting
}
\description{
Labs Regression X
}
\examples{
labsRegressionX(nbaAll)
}

