% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labsRegressionYFunction.R
\name{labsRegressionY}
\alias{labsRegressionY}
\title{Labs Regression Y}
\usage{
labsRegressionY(labModel = NULL, fromReg = NULL)
}
\arguments{
\item{labModel}{model to get Y values from for linear regression fitting}
}
\value{
regression Y vector for fitting labModels
}
\description{
Labs Regression Y
}
\examples{
labsRegressionY(nbaAll)
}

