% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labsZeroProjectionsFunction.R
\name{labZeroProjections}
\alias{labZeroProjections}
\title{Labs Zero Ownership Projections}
\usage{
labZeroProjections(modelDate = "2_12_2017", zeroFit = labZero2Step,
  labsCookies = NULL)
}
\arguments{
\item{modelDate}{date "1_29_2017"}

\item{zeroFit}{regression Model to use for projections}

\item{labsCookies}{labs cookies data for access to models}
}
\value{
df/csv for zero ownership projected players from labs
}
\description{
Labs Zero Ownership Projections
}
\examples{
labZeroProjections("1_29_2017",zeroFit=labZero2Step)
}

