% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaFittingModelsToLoad.R
\name{loadNBAFits}
\alias{loadNBAFits}
\title{NBA Fit Models to Load}
\usage{
loadNBAFits()
}
\value{
loaded models into current session
}
\description{
NBA Fit Models to Load
}
\examples{
loadNBAFits()
}

