% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaModelCorrelationsFunction.R
\name{nbaCorrelations}
\alias{nbaCorrelations}
\title{NBA model Correlations Comparison Function}
\usage{
nbaCorrelations(modelDate = "2_12_2017", model1 = nYearLM1AIC,
  model2 = nYearGLM1, m1Name = "nYearLM1AIC", m2Name = "nYearGLM1",
  fullModel = "nYearLM1AIC_and_nYearGLM1")
}
\arguments{
\item{modelDate}{"1_1_2017"}

\item{model1}{nYearLM1AIC}

\item{model2}{nYearGLM1}

\item{m1Name}{"nYearLM1AIC"}

\item{m2Name}{"nYearGLM1"}

\item{fullModel}{"nYearLM1AIC_and_nYearGLM1"}
}
\value{
df of each models actual Points correlation of projections 20:20
}
\description{
NBA model Correlations Comparison Function
}
\examples{
nba1_1_2017models.Corr <- nbaCorrelations(modelDate="1_1_2017",model1=nYearLM1AIC,model2=nYearGLM1,m1Name="nYearLM1AIC",m2Name="nYearGLM1",fullModel="nYearLM1AIC_and_nYearGLM1")
}

