% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaFitCorrsFunction.R
\name{nbaFitCorrs}
\alias{nbaFitCorrs}
\title{nba Regression Correlations}
\usage{
nbaFitCorrs()
}
\value{
df indicating R^2 correlation value for each fit model
}
\description{
nba Regression Correlations
}
\examples{
nbaFitCorrs()
}

