% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaLinRegFitProjectionsFunction.R
\name{nbaLinRegProjections}
\alias{nbaLinRegProjections}
\title{NBA Linear Regression Model Projections Function}
\usage{
nbaLinRegProjections(modelDate = "11_15_2016", lmFit = nbaFit5_5,
  lmModel = "nbaFit5_5")
}
\arguments{
\item{modelDate}{}

\item{lmFit}{}

\item{lmModel}{}
}
\value{
nba data frame of daily projections using a linear model to predict values
}
\description{
NBA Linear Regression Model Projections Function
}
\examples{
nbaLinRegProjections(modelDate="11_15_2016",lmFit=nbaFit5_5,lmModel = "nbaFit5_5")
}
