% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaMonsterLabProjectionsFunction.R
\name{nbaMonsterLabProjections}
\alias{nbaMonsterLabProjections}
\title{NBA Labs and BBMonster projections Function}
\usage{
nbaMonsterLabProjections(mMonth = "1", mDay = "27", mYear = "2017",
  labsFit = nbaLM12, bbmFit = bbmLMNums1, labsName = "nbaFit12",
  bbmName = "bbmLMNums1", combinedName = "nbaLM12_AND_bbmLMNums1",
  bbmCookie = NULL, errorHandleProp = NULL, errorHandleValue = NULL)
}
\arguments{
\item{mMonth}{"1"}

\item{mDay}{"27"}

\item{mYear}{"2017"}

\item{labsFit}{nbaLM12}

\item{bbmFit}{bbmLMNums1}

\item{labsName}{"nbaLM12"}

\item{bbmName}{"bbmLMNums1"}

\item{combinedName}{"MonsterLab1"}

\item{bbmCookie}{"49390133.1485608109.1.1" the _utmz cookie frm your bbmonster session(update every 30 minutes...)}
}
\value{
df/csv of NBA projections using both models
}
\description{
NBA Labs and BBMonster projections Function
}
\examples{
nbaMonsterLabProjections(mMonth="1",mDay="27",mYear="2017",labsFit=nbaLM12,bbmFit=bbmLMNums1,labsName="nbaFit12",bbmName="bbmLMNums1",combinedName="nbaLM12_AND_bbmLMNums1",bbmCookie="49390133.1485608109.1.1")
}

