% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaOptimal.R
\name{nbaOptimalProj}
\alias{nbaOptimalProj}
\title{NBA Optimal Projections Function}
\usage{
nbaOptimalProj(mMonth = "2", mDay = "12", mYear = "2017",
  labsCookies = NULL, dockerName = "monsterLab", labsFit = allStarLmer,
  labs3k = all3000LM3FitStep, labsName = "allStarLmer",
  labsName3k = "all3000LM3FitStep", bbmFit = bbmLMNums1,
  bbm3k = bbmAll3000LM, bbmName = "bbmLMNums1",
  combinedName = "allStarLmer_AND_bbmLMNums1",
  combined3k = "daily3000Projections", finalName = "OptimalProjections",
  errorHandleProp = NULL, errorHandleValue = NULL)
}
\arguments{
\item{mMonth}{1}

\item{mDay}{30}

\item{mYear}{2017}

\item{labsCookies}{use getLabsCookies();default=NULL}

\item{dockerName}{name of docker container ID to use for getting cookies functions}

\item{labsFit}{m1}

\item{labs3k}{m2}

\item{labsName}{1}

\item{labsName3k}{2}

\item{bbmFit}{3}

\item{bbm3k}{4}

\item{bbmName}{5}

\item{combinedName}{combo}

\item{combined3k}{combo2}

\item{finalName}{final}

\item{errorHandleProp}{fix errors}

\item{errorHandleValue}{fix errors}
}
\value{
names of created df/csv files of daily NBA Projections
}
\description{
NBA Optimal Projections Function
}
\examples{
nbaOptimalProj(mMonth="2",mDay="11",mYear="2017",labsCookies=labsCook,labsFit=nbaLM13,labs3k=all3000LM3FitStep,labsName="nbaFit12",labsName3k="all3000LM3FitStep",bbmFit=bbmLMNums1,bbm3k=bbmAll3000LM,bbmName="bbmLMNums1",combinedName="nbaLM13_AND_bbmLMNums1",combined3k="daily3000Projections",finalName="OptimalProjections",errorHandleProp=NULL,errorHandleValue=NULL)
}

