% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaOptimalProjectionsFunction.R
\name{nbaOptimalProjections}
\alias{nbaOptimalProjections}
\title{Optimal NBA Projections(soFar)}
\usage{
nbaOptimalProjections(mMonth = "1", mDay = "27", mYear = "2017",
  labsFit = nbaLM12, labs3k = all3000LM3FitStep, labsName = "nbaFit12",
  labsName3k = "all3000LM3FitStep", bbmFit = bbmLMNums1,
  bbm3k = bbmAll3000LM, bbmName = "bbmLMNums1",
  combinedName = "nbaLM12_AND_bbmLMNums1",
  combined3k = "daily3000Projections", finalName = "OptimalProjections",
  labsCookie = NULL)
}
\arguments{
\item{mMonth}{month #}

\item{mDay}{day #}

\item{mYear}{"2017"}

\item{labsFit}{Lab Regression}

\item{labs3k}{lab Regression for chumps}

\item{labsName}{name of normal regression}

\item{labsName3k}{name of chump regression}

\item{bbmFit}{bbm regression}

\item{bbm3k}{bbm regression for chumps}

\item{bbmName}{bbm regression name}

\item{combinedName}{nameofBoth}

\item{combined3k}{nameofBoth chump models}

\item{finalName}{name of entire shindig}

\item{labsCookie}{see above for labs}
}
\value{
df/csv of best NBA projections
}
\description{
Optimal NBA Projections(soFar)
}
\examples{
nbaOptimalProjections(mDay="29")
}

