% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaProjectionsFunction_final.R
\name{NbaProjections}
\alias{NbaProjections}
\title{NBA Projections(final)}
\usage{
NbaProjections(modelDate = "3_31_2017", modelSite = "Draftkings",
  lmFit = nba16_17AllLm3Step, lmModel = "nba16_17AllLm3Step",
  errorHandleProp = NULL, errorHandleValue = NULL, cookie = labsC)
}
\arguments{
\item{modelDate}{"3_26_2017"}

\item{modelSite}{"DraftKings"}

\item{lmFit}{lmer10}

\item{lmModel}{"lmer10"}

\item{errorHandleProp}{NULL}

\item{errorHandleValue}{NULL}

\item{cookie}{labsC}
}
\value{
Daily NBA Projections
}
\description{
NBA Projections(final)
}
\examples{
NbaProjections()
}

