% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaTeamSummary.R
\name{nbaTeamSummary}
\alias{nbaTeamSummary}
\title{NBA Team Total Summary}
\usage{
nbaTeamSummary(modelDate = "11_30_2016")
}
\arguments{
\item{modelDate}{}
}
\value{
data frame with relevant Team information for chosen game slate(s)
}
\description{
NBA Team Total Summary
}
\examples{
nbaTeamSummary(modelDate = "11_30_2016")
}

