% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhlProjectionsFunction.R
\name{nhlProjections}
\alias{nhlProjections}
\title{NHL Projections Function}
\usage{
nhlProjections(modelDate = "12_22_2016")
}
\arguments{
\item{modelDate}{}
}
\value{
saves .csv of projections for desired Daily games
}
\description{
NHL Projections Function
}
\examples{
nhlProjections(modelDate)
}
