% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierKDScript.R
\name{outlierKD}
\alias{outlierKD}
\title{OutlierKD Function}
\usage{
outlierKD(dt, var)
}
\arguments{
\item{dt}{data}

\item{var}{variable}
}
\value{
outlier?
}
\description{
OutlierKD Function
}
\examples{
outlierKD(df,variable)
}

