% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegisterParallelPsockCluster.R
\name{parallelPSock}
\alias{parallelPSock}
\title{REGISTER PARALLEL PROCESSOR(PSOCKS)}
\usage{
parallelPSock(cores = 2)
}
\arguments{
\item{cores}{# OF CORES "2"}
}
\value{
NAME OF REGEISTERED PARALLEL PROCESSOR
}
\description{
REGISTER PARALLEL PROCESSOR(PSOCKS)
}
\examples{
parallelPSock(2)
}

