% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionFixingFunction.R
\name{projectionFixing}
\alias{projectionFixing}
\title{Projections formatting for dataFrames from labs for NBA Predictions}
\usage{
projectionFixing(df = labsAll)
}
\arguments{
\item{df}{dataframe}
}
\value{
formatted for use with predict()
}
\description{
Projections formatting for dataFrames from labs for NBA Predictions
}
\examples{
projectionFixing(labsAll)
}

