% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNBACSVsFuntion.R
\name{readNBACSVs}
\alias{readNBACSVs}
\title{read NBA csv saved Models}
\usage{
readNBACSVs(modelDate)
}
\arguments{
\item{modelDate}{}
}
\value{
pulls model into r
}
\description{
read NBA csv saved Models
}
\examples{
nbaModel <- readNBACSVs(modelDate = "2015")
}
