% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNFLCSVsFunction.R
\name{readNFLCSVs}
\alias{readNFLCSVs}
\title{read NFl csv saved Models}
\usage{
readNFLCSVs(modelYear = "2015", modelWeek = NULL, modelId = "554405",
  PositionID = NULL)
}
\arguments{
\item{modelYear}{}

\item{modelWeek}{}

\item{PositionID}{}

\item{mod}{}
}
\value{
pulls model into r
}
\description{
read NFl csv saved Models
}
\examples{
QBM <- readNFLCSVs(modelYear = "2015",modelWeek = NULL ,mod = "", PositionID = NULL)
}
