% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopDockerFunction.R
\name{stopDocker}
\alias{stopDocker}
\title{stop Docker}
\usage{
stopDocker(dID = NULL)
}
\arguments{
\item{dID}{name of running Docker session to stop "distracted_snyder"}
}
\description{
stop Docker
}
\examples{
stopDocker("distracted_snyder")
}

