% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyFantasyNerdProjFormFunction.R
\name{dailyFantasyNerdProjForm}
\alias{dailyFantasyNerdProjForm}
\title{Daily FantasyNerd PRojections Format to csv for upload}
\usage{
dailyFantasyNerdProjForm(modelDate = "12_1_2016", projType = "Averaged")
}
\arguments{
\item{modelDate}{"1_1_2017"}

\item{projType}{name of Projections to be saved as...csv}
}
\value{
df of projections from desired model and day in DFNerd uplaid format
}
\description{
Daily FantasyNerd PRojections Format to csv for upload
}
\examples{
dailyFantasyNerdProjForm(modelDate="12_1_2016",projType="Averaged")
}

