% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLabsCookiesFunction.R
\name{getLabsCookies}
\alias{getLabsCookies}
\title{getLabsCookies}
\usage{
getLabsCookies(dockerName = "labCookies", sport = "mlb")
}
\arguments{
\item{dockerName}{docker container name(can be anything)}

\item{sport}{mlb,nba,nfl,nhl,pga}
}
\value{
all FantasyLab Cookies
}
\description{
getLabsCookies
}
\examples{
getLabsCookies("labCookies","mlb")
}

