% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaModelRetrieve.R
\name{getNBAPlayerModel}
\alias{getNBAPlayerModel}
\title{NBA Model Retrieve}
\usage{
getNBAPlayerModel(modelDate = NULL, modelSite = "Draftkings",
  slate = NULL, exclude = NULL,
  cookie = "__cfduid=d3754702e05d38930dec5c0d04d6ebdad1485624789; LD_U=http\%3A\%2F\%2Fwww.fantasylabs.com\%2F; .AspNet.Cookies=ktnc3djb1JUCI9JKWmXzxZ-F2MHoT9qImiixxGVPd2a2L8JwaiC2Amf5FijhLC1uY8-xsUP3uoRS81HjIG63W2uA0T_FRD5UaovpXg8LyiAOV9Ku60VOC3l6sFCTt1SNFH7hZ_MYgoqUgoucY5lsL9Nm7bmEX79WpB3J9KeCPEXkapD4Ap379jqi1qitYuX1-HQ-r-4PnLy6sJmZSCs7t512GEDlZXYFcvKggIh6Azi2g_DmFJ2y1qzJY2won_-1z4TRdnA84CaJRVGDoYVqa_5oiH_cRSoFVFaWh-PspbXV-WS8TZBjrRMOcw-pjSyad0zP8tVzBFUsI6gpjvWBHaAtDRr2LvDkZamSu_dnRjTj8z7dtPs0N3R7CAdxxe7r-M5jnB6ydRMwGjfJn1yFOypiYPxqeytGrsBAVrFsHR_-vn9M80ZwJv9hUNRPFnBsR7szJuXzpUFJ_Q-xMOnyNPWIfR6gSkqkyMC4kE5Wzd0; flid=5ANbAhdxrkqOWJnJYGArDA; __distillery=13f6068_5d3aa9eb-3e25-403f-be81-5a58b0d371e7-0ef1805cc-ee4b27f5fbc2-cf6a; LD_S=1485637044840; LD_R=; _gat=1; __zlcmid=eogBWbcILnbRCJ; LD_T=b63fdc9b-f2f3-4ff9-cfda-426bb863d4b9; _ga=GA1.2.1049198149.1485624801'")
}
\arguments{
\item{modelDate}{"1_1_2017"}

\item{modelSite}{Draftkings or FanDuel}

\item{slate}{Main,turbo,late,etc}

\item{exclude}{c("ATL","NYK")}

\item{cookie}{for cURL from FLabs (Chrome Browser)..}
}
\value{
Fantasylabs NBA model
}
\description{
NBA Model Retrieve
}
\examples{
getNBAPlayerModel()

}

