% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaProjectionsScript2016_17.R
\name{nbaFullRun}
\alias{nbaFullRun}
\title{NBA Full Run}
\usage{
nbaFullRun(modelDate = "1_1_2017", noStep = nba16_17AllStep,
  step = nYearGLM2, model1name = "nba16_17AllStep",
  model2name = "nYearGLM2", fullModelName = "FULL")
}
\arguments{
\item{modelDate}{"1_1_2017" daily games date}

\item{noStep}{nbaFit5_5 fit model}

\item{step}{lmer10 fit model}
}
\value{
data frame for daily Projections
}
\description{
NBA Full Run
}
\examples{
nbaFullRun(modelDate = "1_1_2017",noStep = nYearLM1AIC,step = nYearGLM1, model1name = "nYearLM1AIC", model2name = "nYearGLM2", fullModelName = "LatestProjFit")
}
