% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaProjectionsScript2016_17.R
\name{nbaFullRun}
\alias{nbaFullRun}
\title{NBA Full Run}
\usage{
nbaFullRun(modelDate = "12_1_2016", noStep = nbaFit5_5,
  step = nbaFit5_5Step)
}
\arguments{
\item{modelDate}{}

\item{noStep}{}

\item{step}{}
}
\value{
data frame for daily Projections
}
\description{
NBA Full Run
}
\examples{
nbaFullRun(modelDate = "12_1_2016",noStep = nbaFit5_5,step = nbaFit5_5Step)
}
