% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaProjectionsScript2016_17.R
\name{nbaFullRun}
\alias{nbaFullRun}
\title{NBA Full Run}
\usage{
nbaFullRun(modelDate = "12_1_2016", noStep = nbaFit5_5, step = lmer10,
  model1name = "nbaFit5_5", model2name = "lmer10",
  fullModelName = "Averaged")
}
\arguments{
\item{modelDate}{"1_1_2017" daily games date}

\item{noStep}{nbaFit5_5 fit model}

\item{step}{lmer10 fit model}
}
\value{
data frame for daily Projections
}
\description{
NBA Full Run
}
\examples{
nbaFullRun(modelDate = "12_1_2016",noStep = nbaFit5_5,step = nbaFit5_5Step, model1name = "nbaFit5_5", model2name = "lmer10", fullModelName = "Averaged")
}
