% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaLinRegFitProjectionsFunction.R
\name{nbaLinRegProjections}
\alias{nbaLinRegProjections}
\title{NBA linear regression fit model projections function}
\usage{
nbaLinRegProjections(modelDate = "11_15_2016", lmFit = nbaFit5_5,
  lmModel = "nbaFit5_5")
}
\arguments{
\item{lmModel}{}
}
\value{
nba data frame Daily projections
}
\description{
NBA linear regression fit model projections function
}
\examples{
nbaLinRegFitProjections(modelDate = "11_15_2016")
}
