% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNFLCSVsFunction.R
\name{readNFLCSVs}
\alias{readNFLCSVs}
\title{Read NFl csv saved Models}
\usage{
readNFLCSVs(modelYear = "2015", modelSite = "Draftkings",
  modelWeek = NULL, modelId = NULL, PositionID = NULL)
}
\arguments{
\item{modelYear}{"2016"}

\item{modelSite}{Draftkings or Fanduel}

\item{modelWeek}{17}

\item{modelId}{FantasyLabs modelID(Part of name of saved csv file to retrieve)}

\item{PositionID}{"QB"}
}
\value{
pulls model into r
}
\description{
Read NFl csv saved Models
}
\examples{
QBM <- readNFLCSVs(modelYear = "2015",modelSite="Draftkings",modelWeek = NULL ,mod = "", PositionID = NULL)
}
