% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNFLCSVsFunction.R
\name{readNFLCSVs}
\alias{readNFLCSVs}
\title{Read NFl csv saved Models}
\usage{
readNFLCSVs(modelYear = "2015", modelWeek = NULL, modelId = NULL,
  PositionID = NULL)
}
\arguments{
\item{modelYear}{"2016"}

\item{modelWeek}{17}

\item{PositionID}{"QB"}

\item{mod}{NULL}
}
\value{
pulls model into r
}
\description{
Read NFl csv saved Models
}
\examples{
QBM <- readNFLCSVs(modelYear = "2015",modelWeek = NULL ,mod = "", PositionID = NULL)
}
