% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNHLCsvsFunction.R
\name{readNHLCSVs}
\alias{readNHLCSVs}
\title{read NHL csv saved Models}
\usage{
readNHLCSVs(modelDate = "12_23_2016", modelPosition = "All")
}
\arguments{
\item{modelPosition}{}
}
\value{
pulls model into r
}
\description{
read NHL csv saved Models
}
\examples{
nhl12_23_2016 <- readNHLCSVs(modelDate = "12_23_2016",modelPosition="All")
}
